streaming/producer.py
import os
import json
from kafka import KafkaProducer
from dotenv import load_dotenv

load_dotenv()

KAFKA_BROKER = os.getenv('KAFKA_BROKER', 'localhost:9092')
producer = KafkaProducer(
    bootstrap_servers=KAFKA_BROKER,
    value_serializer=lambda v: json.dumps(v).encode('utf-8')
)

def send_transaction(chain, tx_data):
    topic = f'{chain}_transactions'
    producer.send(topic, tx_data)
    producer.flush() 